==> notebooks/04_generate_stats.ipynb <==
import pandas as pd
from scipy import stats
import numpy as np
import json
from sharedcontrolpaper.force_sensitive_stopping_task_utils import print_means_t_test, print_effect_size_and_ci, calc_stats_ind, convert_formats_back, rename_index_column

## Load data from JSON files

with open('force_sensitive_data.json', 'r') as f:
    loaded_data = json.load(f)

force_sensitive_stopping_task_ssrt = pd.DataFrame.from_records(loaded_data['force_sensitive_stopping_task_ssrt'])
go_task_accuracy_before_stop_onset = pd.DataFrame.from_records(loaded_data['go_task_accuracy_before_stop_onset'])
go_task_accuracy_after_stop_onset = pd.DataFrame.from_records(loaded_data['go_task_accuracy_after_stop_onset'])
duration_of_inhibition = pd.DataFrame.from_records(loaded_data['duration_of_inhibition'])
ssrt_first_half = pd.DataFrame.from_records(loaded_data['ssrt_first_half'])
ssrt_second_half = pd.DataFrame.from_records(loaded_data['ssrt_second_half'])
shared_control_metrics = convert_formats_back(loaded_data['shared_control_metrics'])
first_non_zero_pressure_timestamp = pd.DataFrame.from_records(loaded_data['first_non_zero_pressure_timestamp'])
first_full_pressure_timestamp = pd.DataFrame.from_records(loaded_data['first_full_pressure_timestamp'])

with open('simple_stop_data.json', 'r') as f:
    loaded_data = json.load(f)

simple_stop_metrics = pd.DataFrame.from_records(loaded_data['simple_stop_metrics'])

with open('ai_survey_data.json', 'r') as f:
    loaded_data = json.load(f)

survey_scores = convert_formats_back(loaded_data['survey_scores'])


### Remove the Mean and SD rows for Simple Stop

simple_stop_ssrt = simple_stop_metrics.iloc[:-2]

### Merge Simple Stop and Force Sensitive Stopping Task SSRTs

merged_df = pd.merge(force_sensitive_stopping_task_ssrt, simple_stop_metrics[['ssrt', 'ssrt_without_short_ssd_trials', 'ssrt_without_short_ssd_subs']], 
                    left_index=True, right_index=True, how='left')
merged_df.rename(columns={
    'ssrt': 'simple_stop_ssrt',
    'ssrt_without_short_ssd_trials': 'simple_stop_ssrt_without_short_ssd_trials',
    'ssrt_without_short_ssd_subs': 'simple_stop_ssrt_without_short_ssd_subs'
}, inplace=True)

rename_index_column(merged_df)

## Planned Statistical Tests

### AI-failed vs Non-AI SSRT

print_means_t_test(merged_df, 'ai_failed', 'non_ai')

print_effect_size_and_ci(merged_df,'ai_failed', 'non_ai')

### AI-assisted vs AI-failed SSRT

print_means_t_test(merged_df, 'ai_assisted', 'ai_failed')

print_effect_size_and_ci(merged_df, 'ai_assisted', 'ai_failed')

### Non-AI SSRT vs Simple Stop SSRT

print_means_t_test(merged_df, 'non_ai', 'simple_stop_ssrt')
non_ai_vs_simple_corr = np.corrcoef(merged_df['non_ai'], merged_df['simple_stop_ssrt'])[1][0]
print(f"Correlation: {non_ai_vs_simple_corr:.2f}")

print_effect_size_and_ci(merged_df, 'non_ai', 'simple_stop_ssrt')

## Order Effects

### Calculate order effects between subjects who had the AI block first vs. the Non-AI block first

# Initialize subjects who had the Non-AI and AI Blocks first
non_ai_first_subs = ['s004', 's009', 's008', 's011', 's012', 's015', 's016', 's019', 's020', 's023', 's024', 's027', 's028', 's031', 's032', 's035', 's036', 's039', 's040', 's043']
ai_first_subs = ['s005', 's006', 's007', 's010', 's013', 's014', 's017', 's018', 's021', 's022', 's025', 's026', 's029', 's030', 's033', 's034', 's037', 's038', 's041', 's042']

ssrt_non_ai_first_non_ai = force_sensitive_stopping_task_ssrt.loc[non_ai_first_subs, "non_ai"].values
ssrt_non_ai_first_ai_failed = force_sensitive_stopping_task_ssrt.loc[non_ai_first_subs, "ai_failed"].values
ssrt_ai_first_non_ai = force_sensitive_stopping_task_ssrt.loc[ai_first_subs, "non_ai"].values
ssrt_ai_first_ai_failed = force_sensitive_stopping_task_ssrt.loc[ai_first_subs, "ai_failed"].values

# Calculate differences
diff_non_ai_first = ssrt_non_ai_first_ai_failed - ssrt_non_ai_first_non_ai
diff_ai_first = ssrt_ai_first_ai_failed - ssrt_ai_first_non_ai

# Create DataFrame
diff_df = pd.DataFrame({
    'Non-AI First': diff_non_ai_first,
    'AI First': diff_ai_first,
})

print_means_t_test(diff_df, 'AI First', 'Non-AI First')
print_effect_size_and_ci(diff_df, 'AI First', 'Non-AI First')

### Calculate AI-failed vs. Non-AI SSRT depending on whether they occurred in the first half or second half of the block

#### First half

print_means_t_test(ssrt_first_half, 'ai_failed', 'non_ai')
print_effect_size_and_ci(ssrt_first_half, 'ai_failed', 'non_ai')

#### Second Half

print_means_t_test(ssrt_second_half, 'ai_failed', 'non_ai')
print_effect_size_and_ci(ssrt_second_half, 'ai_failed', 'non_ai')

## Additional T-tests and Confidence Intervals

### Non-AI vs AI-failed Duration of Inhibition

print_means_t_test(duration_of_inhibition, 'ai_failed', 'non_ai')
print_effect_size_and_ci(duration_of_inhibition, 'ai_failed', 'non_ai')

### Non-AI vs AI-failed Go Task Accuracy Before Stop Onset

print_means_t_test(go_task_accuracy_before_stop_onset, 'ai_failed', 'non_ai')
print_effect_size_and_ci(go_task_accuracy_before_stop_onset, 'ai_failed', 'non_ai')

### Non-AI vs AI-failed Go Task Accuracy After Stop Onset (aka Stop Success Rate)

print_means_t_test(go_task_accuracy_after_stop_onset, 'non_ai', 'ai_failed')
print_effect_size_and_ci(go_task_accuracy_after_stop_onset, 'non_ai', 'ai_failed')

### Non-AI vs AI-failed SSRT in subjects who did not show proactive slowing

non_proactive_slowing_subs = go_task_accuracy_before_stop_onset[
    go_task_accuracy_before_stop_onset['non_ai'] > go_task_accuracy_before_stop_onset['ai_failed']
].index.tolist()

ssrt_data = force_sensitive_stopping_task_ssrt.copy() # Avoid modifying original dataframe
ssrt_data = ssrt_data.rename_axis('subject_id').reset_index() #add subject_id column
ssrt_data = ssrt_data[ssrt_data['subject_id'] != 'mean'] #remove mean row

# Calculate difference and assign slowing type using vectorized operations
ssrt_data['difference_ms'] = ssrt_data['ai_failed'] - ssrt_data['non_ai']
ssrt_data['slowing_type'] = np.where(ssrt_data['subject_id'].isin(non_proactive_slowing_subs), 'Non-Proactive', 'Proactive')

# Filter for non-proactive slowing data
non_proactive_slowing_data = ssrt_data[ssrt_data['slowing_type'] == 'Non-Proactive']

print_means_t_test(non_proactive_slowing_data, 'ai_failed', 'non_ai')
print_effect_size_and_ci(non_proactive_slowing_data, 'ai_failed', 'non_ai')

### T-test comparing AI-failed minus Non-AI SSRT in subjects who proactively slowed and did not proactively slow

non_proactive_differences = ssrt_data[ssrt_data['slowing_type'] == 'Non-Proactive']['difference_ms'].dropna()
proactive_differences = ssrt_data[ssrt_data['slowing_type'] == 'Proactive']['difference_ms'].dropna()

calc_stats_ind(non_proactive_differences, proactive_differences)

print(f"Mean difference (Non-Proactive): {ssrt_data[ssrt_data['slowing_type'] == 'Non-Proactive']['difference_ms'].mean():.2f} ms")
print(f"Mean difference (Proactive): {ssrt_data[ssrt_data['slowing_type'] == 'Proactive']['difference_ms'].mean():.2f} ms")

### T-test comparing the first non-zero pressure timestamp between the AI and Non-AI blocks

print_means_t_test(first_non_zero_pressure_timestamp, 'non_ai', 'ai_failed') #Both AI-failed and AI-assisted columns have the same values
print_effect_size_and_ci(first_non_zero_pressure_timestamp, 'non_ai', 'ai_failed')

### T-test comparing the first full pressure timestamp between the AI and Non-AI blocks

print_means_t_test(first_full_pressure_timestamp, 'non_ai', 'ai_failed') #Both AI-failed and AI-assisted columns have the same values
print_effect_size_and_ci(first_full_pressure_timestamp, 'non_ai', 'ai_failed')

## Correlate survey scores with SSRT

survey_scores_df = pd.DataFrame(survey_scores)
avg_df = merged_df.merge(survey_scores_df, on='subject_id', how='left')
avg_df['difference_ai_failed_and_non_ai_ssrt'] = avg_df['ai_failed'] - avg_df['non_ai']
correlation, pval = stats.pearsonr(avg_df['average_score'], avg_df["difference_ai_failed_and_non_ai_ssrt"])
print(f"Correlation and p-value between SSRT and survey scores: {correlation:.2f}, {pval:.3f}")

## Violations of Context Independence in Simple Stop Data

### a) Non-AI SSRT vs Simple Stop SSRT Excluding trials with SSD < 200 ms

print_means_t_test(merged_df, 'non_ai', 'simple_stop_ssrt_without_short_ssd_trials')
print_effect_size_and_ci(merged_df, 'non_ai', 'simple_stop_ssrt_without_short_ssd_trials')

### b) Non-AI SSRT vs Simple Stop SSRT Excluding subjects with average SSDs < 200ms

# Remove rows with NaN in simple_stop_ssrt_without_short_ssd_subs
filtered_df = merged_df.dropna(subset=['simple_stop_ssrt_without_short_ssd_subs'])

print_means_t_test(filtered_df, 'non_ai', 'simple_stop_ssrt_without_short_ssd_subs')
print_effect_size_and_ci(filtered_df, 'non_ai', 'simple_stop_ssrt_without_short_ssd_subs')

## Store merged_df to be accessed later

data_to_save = {
    'merged_df': merged_df.to_dict(),
}

with open('merged_df.json', 'w') as f:
    json.dump(data_to_save, f, indent=4)
